+++ scripts/logtalk_backend_select.sh
@@ -3,7 +3,7 @@
 #############################################################################
 ## 
 ##   Logtalk back-end Prolog compiler select script
-##   Last updated on May 16, 2018
+##   Last updated on May 17, 2018
 ## 
 ##   This file is part of Logtalk <https://logtalk.org/>  
 ##   Copyright 1998-2018 Paulo Moura <pmoura@logtalk.org>
@@ -24,48 +24,48 @@
 
 
 print_version() {
-	echo "Current $(basename $0) version:"
-	echo "  0.7"
+	echo "Current $(basename "$0") version:"
+	echo "  0.8"
 	exit 0
 }
 
 
 list_backends() {
     echo "Available back-end Prolog compilers:"
-	if [ -e $(command -v bplgt) ] && [ "$(command -v bp)" != "" ] ; then
+	if [ -e "$(command -v bplgt)" ] && [ "$(command -v bp)" != "" ] ; then
 		echo -n "  bplgt"
 	fi
-	if [ -e $(command -v cxlgt) ] && [ "$(command -v cxprolog)" != "" ] ; then
+	if [ -e "$(command -v cxlgt)" ] && [ "$(command -v cxprolog)" != "" ] ; then
 		echo -n "  cxlgt"
 	fi
-	if [ -e $(command -v eclipselgt) ] && [ "$(command -v eclipse)" != "" ] ; then
+	if [ -e "$(command -v eclipselgt)" ] && [ "$(command -v eclipse)" != "" ] ; then
 		echo -n "  eclipselgt"
 	fi
-	if [ -e $(command -v gplgt) ] && [ "$(command -v gprolog)" != "" ] ; then
+	if [ -e "$(command -v gplgt)" ] && [ "$(command -v gprolog)" != "" ] ; then
 		echo -n "  gplgt"
 	fi
-	if [ -e $(command -v jiplgt) ] && [ "$(command -v jipconsole.sh)" != "" ] ; then
+	if [ -e "$(command -v jiplgt)" ] && [ "$(command -v jipconsole.sh)" != "" ] ; then
 		echo -n "  jiplgt"
 	fi
-	if [ -e $(command -v lplgt) ] && [ "$(command -v lprolog)" != "" ] ; then
+	if [ -e "$(command -v lplgt)" ] && [ "$(command -v lprolog)" != "" ] ; then
 		echo -n "  lplgt"
 	fi
-	if [ -e $(command -v qplgt) ] && [ "$(command -v qp)" != "" ] ; then
+	if [ -e "$(command -v qplgt)" ] && [ "$(command -v qp)" != "" ] ; then
 		echo -n "  qplgt"
 	fi
-	if [ -e $(command -v sicstuslgt) ] && [ "$(command -v sicstus)" != "" ] ; then
+	if [ -e "$(command -v sicstuslgt)" ] && [ "$(command -v sicstus)" != "" ] ; then
 		echo -n "  sicstuslgt"
 	fi
-	if [ -e $(command -v swilgt) ] && [ "$(command -v swipl)" != "" ] ; then
+	if [ -e "$(command -v swilgt)" ] && [ "$(command -v swipl)" != "" ] ; then
 		echo -n "  swilgt"
 	fi
-	if [ -e $(command -v xsblgt) ] && [ "$(command -v xsb)" != "" ] ; then
+	if [ -e "$(command -v xsblgt)" ] && [ "$(command -v xsb)" != "" ] ; then
 		echo -n "  xsblgt"
 	fi
-	if [ -e $(command -v xsbmtlgt) ] && [ "$(command -v xsb-mt)" != "" ] ; then
+	if [ -e "$(command -v xsbmtlgt)" ] && [ "$(command -v xsb-mt)" != "" ] ; then
 		echo -n "  xsbmtlgt"
 	fi
-	if [ -e $(command -v yaplgt) ] && [ "$(command -v yap)" != "" ] ; then
+	if [ -e "$(command -v yaplgt)" ] && [ "$(command -v yap)" != "" ] ; then
 		echo -n "  yaplgt"
 	fi
 	echo
@@ -74,13 +74,29 @@ list_backends() {
 
 
 show_selected() {
-    echo "Current Prolog integration script:"
-    if [ -e $(command -v logtalk) ] && [ "$(command -v logtalk)" != "" ] ; then
+	echo "Current Prolog integration script:"
+	if [ -e "$(command -v logtalk)" ] && [ "$(command -v logtalk)" != "" ] ; then
 		echo -n "  "
-		readlink $(command -v logtalk)
-    else
-        echo "  none"
-    fi
+		readlink "$(command -v logtalk)"
+	else
+		echo "  none"
+	fi
+	exit 0
+}
+
+
+remove_link() {
+	if [ -e "$(command -v logtalk)" ] && [ "$(command -v logtalk)" != "" ] ; then
+		if rm "$(command -v logtalk)" ; then
+			echo "Symbolic link removed"
+		else
+			echo "An error occurred while trying to remove the symbolic link!" >&2
+			echo "Check that you are executing this script with the necessary permissions." >&2
+			exit 1
+		fi
+	else
+		echo "Symbolic link not defined"
+	fi
 	exit 0
 }
 
@@ -95,41 +111,43 @@ usage_help() {
 	echo "  $(basename "$0") -v"
 	echo "  $(basename "$0") -l"
 	echo "  $(basename "$0") -s"
+	echo "  $(basename "$0") -r"
 	echo "  $(basename "$0") -h"
 	echo
 	echo "Optional arguments:"
 	echo "  -v print script version"
 	echo "  -l list available Prolog integration scripts"
 	echo "  -s show the currently selected Prolog integration script"
+	echo "  -r remove the \"logtalk\" symbolic link"
 	echo "  -h help"
 	echo
 }
 
 
 valid_backend() {
-	if [ "$1" == "bplgt" ] && [ -e $(command -v bplgt) ]  && [ "$(command -v bp)" != "" ] ; then
+	if [ "$1" == "bplgt" ] && [ -e "$(command -v bplgt)" ]  && [ "$(command -v bp)" != "" ] ; then
 		return 0
-	elif [ "$1" == "cxlgt" ] && [ -e $(command -v cxlgt) ]  && [ "$(command -v cxprolog)" != "" ] ; then
+	elif [ "$1" == "cxlgt" ] && [ -e "$(command -v cxlgt)" ]  && [ "$(command -v cxprolog)" != "" ] ; then
 		return 0
-	elif [ "$1" == "eclipselgt" ] && [ -e $(command -v eclipselgt) ]  && [ "$(command -v eclipse)" != "" ] ; then
+	elif [ "$1" == "eclipselgt" ] && [ -e "$(command -v eclipselgt)" ]  && [ "$(command -v eclipse)" != "" ] ; then
 		return 0
-	elif [ "$1" == "gplgt" ] && [ -e $(command -v gplgt) ]  && [ "$(command -v gprolog)" != "" ] ; then
+	elif [ "$1" == "gplgt" ] && [ -e "$(command -v gplgt)" ]  && [ "$(command -v gprolog)" != "" ] ; then
 		return 0
-	elif [ "$1" == "jiplgt" ] && [ -e $(command -v jiplgt) ]  && [ "$(command -v jipconsole.sh)" != "" ] ; then
+	elif [ "$1" == "jiplgt" ] && [ -e "$(command -v jiplgt)" ]  && [ "$(command -v jipconsole.sh)" != "" ] ; then
 		return 0
-	elif [ "$1" == "lplgt" ] && [ -e $(command -v lplgt) ]  && [ "$(command -v lprolog)" != "" ] ; then
+	elif [ "$1" == "lplgt" ] && [ -e "$(command -v lplgt)" ]  && [ "$(command -v lprolog)" != "" ] ; then
 		return 0
-	elif [ "$1" == "qplgt" ] && [ -e $(command -v qplgt) ]  && [ "$(command -v qp)" != "" ] ; then
+	elif [ "$1" == "qplgt" ] && [ -e "$(command -v qplgt)" ]  && [ "$(command -v qp)" != "" ] ; then
 		return 0
-	elif [ "$1" == "sicstuslgt" ] && [ -e $(command -v sicstuslgt) ]  && [ "$(command -v sicstus)" != "" ] ; then
+	elif [ "$1" == "sicstuslgt" ] && [ -e "$(command -v sicstuslgt)" ]  && [ "$(command -v sicstus)" != "" ] ; then
 		return 0
-	elif [ "$1" == "swilgt" ] && [ -e $(command -v swilgt) ]  && [ "$(command -v swipl)" != "" ] ; then
+	elif [ "$1" == "swilgt" ] && [ -e "$(command -v swilgt)" ]  && [ "$(command -v swipl)" != "" ] ; then
 		return 0
-	elif [ "$1" == "xsblgt" ] && [ -e $(command -v xsblgt) ]  && [ "$(command -v xsb)" != "" ] ; then
+	elif [ "$1" == "xsblgt" ] && [ -e "$(command -v xsblgt)" ]  && [ "$(command -v xsb)" != "" ] ; then
 		return 0
-	elif [ "$1" == "xsbmtlgt" ] && [ -e $(command -v xsbmtlgt) ]  && [ "$(command -v xsb-mt)" != "" ] ; then
+	elif [ "$1" == "xsbmtlgt" ] && [ -e "$(command -v xsbmtlgt)" ]  && [ "$(command -v xsb-mt)" != "" ] ; then
 		return 0
-	elif [ "$1" == "yaplgt" ] && [ -e $(command -v yaplgt) ]  && [ "$(command -v yap)" != "" ] ; then
+	elif [ "$1" == "yaplgt" ] && [ -e "$(command -v yaplgt)" ]  && [ "$(command -v yap)" != "" ] ; then
 		return 0
 	else
 		return 1
@@ -138,34 +156,32 @@ valid_backend() {
 
 
 switch_backend() {
-	valid_backend "$1"
-	if [ 0 != ${?} ]; then
-    	echo "Invalid Prolog integration script: $1" >&2
-    	exit 1
-	else
-		cd $(dirname $(command -v "$1")) || exit 1
+	if valid_backend "$1" ; then
+		cd "$(dirname "$(command -v "$1")")" || exit 1
 		rm -f logtalk
-		ln -sf "$1" logtalk
-		error=$?
-		if [ 0 != $error ]; then
-			echo "An error occurred when switching the default Prolog integration script!" >&2
-			echo "Check that you are executing this script with the necessary permissions." >&2
-			exit 1
-    	else
+		if ln -sf "$1" logtalk ; then
 			echo "Switched to the Prolog integration script:"
 			echo "  $1"
 			exit 0
+		else
+			echo "An error occurred when switching the default Prolog integration script!" >&2
+			echo "Check that you are executing this script with the necessary permissions." >&2
+			exit 1
 		fi
+	else
+		echo "Invalid Prolog integration script: $1" >&2
+		exit 1
 	fi
 }
 
 
-while getopts "vlsh" option
+while getopts "vlsrh" option
 do
 	case $option in
 		v) print_version;;
 		l) list_backends;;
 		s) show_selected;;
+		r) remove_link;;
 		h) usage_help; exit;;
 		*) usage_help; exit;;
 	esac
@@ -176,11 +192,10 @@ if [ "$1" == "" ]; then
 	usage_help
 	exit 0
 else
-	switch_backend "$1"
-	error=$?
-	if [ 0 != $error ]; then
+	if switch_backend "$1" ; then
+		exit 0
+	else
 		echo "An error occurred when switching to the $1 Prolog integration script!" >&2
 		exit 1
 	fi
-	exit 0
 fi